% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RGF_cleanup_temp_files}
\alias{RGF_cleanup_temp_files}
\title{Delete all temporary files of the created RGF estimators}
\usage{
RGF_cleanup_temp_files()
}
\description{
Delete all temporary files of the created RGF estimators
}
\details{
This function deletes all temporary files of the created RGF estimators. See the issue \emph{https://github.com/RGF-team/rgf/issues/75} for more details.
}
\examples{

\dontrun{
library(RGF)

RGF_cleanup_temp_files()
}
}
\references{
\emph{https://github.com/RGF-team/rgf/tree/master/python-package}
}
