% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_class.R
\name{Internal_class}
\alias{Internal_class}
\title{Internal R6 class for all secondary functions used in RGF and FastRGF}
\description{
Internal R6 class for all secondary functions used in RGF and FastRGF

Internal R6 class for all secondary functions used in RGF and FastRGF
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fit}{\code{Internal_class$fit()}}
\item \href{#method-predict}{\code{Internal_class$predict()}}
\item \href{#method-predict_proba}{\code{Internal_class$predict_proba()}}
\item \href{#method-cleanup}{\code{Internal_class$cleanup()}}
\item \href{#method-get_params}{\code{Internal_class$get_params()}}
\item \href{#method-score}{\code{Internal_class$score()}}
\item \href{#method-feature_importances}{\code{Internal_class$feature_importances()}}
\item \href{#method-dump_model}{\code{Internal_class$dump_model()}}
\item \href{#method-save_model}{\code{Internal_class$save_model()}}
\item \href{#method-clone}{\code{Internal_class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Internal_class$fit(x, y, sample_weight = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Internal_class$predict(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_proba"></a>}}
\subsection{Method \code{predict_proba()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Internal_class$predict_proba(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleanup"></a>}}
\subsection{Method \code{cleanup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Internal_class$cleanup()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_params"></a>}}
\subsection{Method \code{get_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Internal_class$get_params(deep = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-score"></a>}}
\subsection{Method \code{score()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Internal_class$score(x, y, sample_weight = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feature_importances"></a>}}
\subsection{Method \code{feature_importances()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Internal_class$feature_importances()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dump_model"></a>}}
\subsection{Method \code{dump_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Internal_class$dump_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_model"></a>}}
\subsection{Method \code{save_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Internal_class$save_model(filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Internal_class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
