% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{Internal_class}
\alias{Internal_class}
\title{Internal R6 class for all secondary functions used in RGF and FastRGF}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Internal_class
}
\description{
Internal R6 class for all secondary functions used in RGF and FastRGF
}
\keyword{internal}
